style: remove duplicate import
