+++ pdf_chunker/adapters/__init__.py
@@ -1,4 +1,4 @@
-from . import emit_jsonl, io_pdf
+from . import ai_enrich, emit_jsonl, io_pdf
 
 try:  # pragma: no cover - optional dependency
     from . import io_epub  # type: ignore
@@ -7,5 +7,5 @@ except ModuleNotFoundError:  # pragma: no cover
 
 from .io_epub import describe_epub, read_epub  # noqa: F401
 
-__all__ = ["emit_jsonl", "io_epub", "io_pdf"]
+__all__ = ["ai_enrich", "emit_jsonl", "io_epub", "io_pdf"]
 __all__ += ["describe_epub", "read_epub"]
